chore: Move ResultUtil to domain layer
