+++ data/src/main/java/in/koreatech/koin/data/repository/PreSignedUrlRepositoryImpl.kt
@@ -3,8 +3,8 @@ package `in`.koreatech.koin.data.repository
 import `in`.koreatech.koin.data.requestbody.S3RequestBody
 import `in`.koreatech.koin.data.source.local.UploadImageLocalDataSource
 import `in`.koreatech.koin.data.source.remote.PreSignedUrlRemoteDataSource
-import `in`.koreatech.koin.data.util.suspendRunCatching
 import `in`.koreatech.koin.domain.repository.PreSignedUrlRepository
+import `in`.koreatech.koin.domain.util.suspendRunCatching
 import java.io.InputStream
 import javax.inject.Inject
 import okhttp3.MediaType.Companion.toMediaType
