--- data/src/main/java/in/koreatech/koin/data/util/ResultUtil.kt
@@ -1,13 +1,0 @@
-package `in`.koreatech.koin.data.util
-
-import kotlin.coroutines.cancellation.CancellationException
-
-suspend fun <T> suspendRunCatching(block: suspend () -> T): Result<T> {
-    return try {
-        Result.success(block())
-    } catch (e: CancellationException) {
-        throw e
-    } catch (e: Exception) {
-        Result.failure(e)
-    }
-}
